* possible? solution on bad amount format
Some times appears an error (SIS0019) when user is moved to payment
page. SIS0019 error indicate that the data is not sending on well
format.

Looking deeper, when error appears amount is sending as 26014.999999999996
(in example), when it always should be an integer value (float amount
should be multiplied by 100).

Seems to be an error on how amount is processed, so i change this
behaviour to ensure? that. Even that, i not really sure why is this
malfunction on first time, so maybe this do nothing :-)